+++ RELEASENOTES.md
@@ -1,1 +1,2 @@
 - When blob storage credentials are provided to the CLI but will not be used for a GHES migration, log a clear warning, not an info message
+- Improve the error thrown when you have insufficient permissions for the target GitHub organization to explicitly mention the relevant organization